requirements.txt
Django==5.2.11
djangorestframework
psutil
